// configurator_frontend/src/payload-types.ts
export interface Media {
  id: string;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    admin_thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface ManualShowAction {
  nodes?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualShow';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ManualHideAction".
 */
export interface ManualHideAction {
  nodes?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualHide';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ManualSetColorAction".
 */
export interface ManualSetColorAction {
  materials?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  colorCode: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualSetColor';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "SetUserImageAction".
 */
export interface SetUserImageAction {
  width: number;
  height: number;
  material: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'setUserImage';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "SetTextureAction".
 */
export interface SetTextureAction {
  materials?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  wrapS: 'repeat' | 'mirroredRepeat' | 'clampToEdge';
  wrapT: 'repeat' | 'mirroredRepeat' | 'clampToEdge';
  repeatX: number;
  repeatY: number;
  offsetX: number;
  offsetY: number;
  id?: string | null;
  blockName?: string | null;
  blockType: 'setTexture';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "SetColorAction".
 */
export interface SetColorAction {
  materials?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'setColorAction';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ShowPartAction".
 */
export interface ShowPartAction {
  id?: string | null;
  blockName?: string | null;
  blockType: 'showPart';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ManualSetTextureAction".
 */
export interface ManualSetTextureAction {
  materials?:
    | {
        name: string;
        id?: string | null;
      }[]
    | null;
  colorMap: string | Media;
  color?: string | null;
  aoMap?: string | Media | null;
  aoIntensity?: number | null;
  displacementMap?: string | Media | null;
  displacementScale?: number | null;
  normalMap?: string | Media | null;
  normalScaleX?: number | null;
  normalScaleY?: number | null;
  roughnessMap?: string | Media | null;
  roughness?: number | null;
  metalnessMap?: string | Media | null;
  metalness?: number | null;
  bumpMap?: string | Media | null;
  bumpScale?: number | null;
  emissiveMap?: string | Media | null;
  emissiveColor?: string | null;
  emissiveIntensity?: number | null;
  wrapS: 'repeat' | 'mirroredRepeat' | 'clampToEdge';
  wrapT: 'repeat' | 'mirroredRepeat' | 'clampToEdge';
  repeatX: number;
  repeatY: number;
  offsetX: number;
  offsetY: number;
  id?: string | null;
  blockName?: string | null;
  blockType: 'manualSetTextureAction';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "UnsetAttributeAction".
 */
export interface UnsetAttributeAction {
  attributeCode: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'unsetAttributeAction';
}
// ...
export interface ColorAttribute {
  name: string;
  code: string;
  colors?: (string | Color)[] | null;
  customColors?: boolean | null;
  defaultValue?: (string | null) | Color;
  actions?:
    | (
        | ManualShowAction
        | ManualHideAction
        | ManualSetColorAction
        | SetUserImageAction
        | SetTextureAction
        | SetColorAction
        | ShowPartAction
        | ManualSetTextureAction
        | UnsetAttributeAction
      )[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'colorAttribute';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "colors".
 */
export interface Color {
  id: string;
  name: string;
  colorCode: string;
  updatedAt: string;
  createdAt: string;
}